# OpenAPI.yaml
openapi: 3.0.1
info:
  title: API Vulpes
  description: ""
  contact:
    name: Kaique de Miranda
    email: kbdemiranda@hotmail.com
servers:
- url: http://127.0.0.1:8080/api
  description: Generated server url
...
paths:
...
  /assinantes/{id}:
    get:
      tags:
      - Assinantes
      summary: Buscar assinante
      description: Retorna um assinante pelo ID
      operationId: buscarAssinante
      parameters:
      - name: id
        in: path
        description: ID do assinante
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "404":
          description: Assinante não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Assinante encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Assinante'
      security:
      - bearer-key: []
    put:
      tags:
      - Assinantes
      summary: Atualizar assinante
      description: Atualiza dados de um assinante existente
      operationId: atualizarAssinante
      parameters:
      - name: id
        in: path
        description: ID do assinante
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Assinante'
        required: true
      responses:
        "201":
          description: Assinante atualizado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Assinante'
        "404":
          description: Assinante não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
    delete:
      tags:
      - Assinantes
      summary: Excluir assinante
      description: Remove um assinante pelo ID
      operationId: excluirAssinante
      parameters:
      - name: id
        in: path
        description: ID do assinante
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "404":
          description: Assinante não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "204":
          description: Assinante excluído
      security:
      - bearer-key: []
...
  /assinantes:
    get:
      tags:
      - Assinantes
      summary: Listar assinantes
      description: "Retorna lista paginada de assinantes, com filtro opcional por\
        \ nome"
      operationId: listarAssinantes
      parameters:
      - name: nome
        in: query
        description: Filtro por nome
        required: false
        schema:
          type: string
        example: João
      - name: pagina
        in: query
        description: Número da página (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
        example: 0
      - name: quantidade
        in: query
        description: Quantidade por página
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
        example: 10
      responses:
        "500":
          description: Erro interno
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Lista retornada com sucesso
          content:
            '*/*':
              schema:
                type: object
        "401":
          description: Não autorizado
      security:
      - bearer-key: []
    post:
      tags:
      - Assinantes
      summary: Cadastrar assinante
      description: Cria um novo assinante
      operationId: cadastrarAssinante
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Assinante'
        required: true
      responses:
        "201":
          description: Assinante criado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Assinante'
        "400":
          description: Dados inválidos
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
  /assinantes/{id}/associar-plataformas:
    post:
      tags:
      - Assinantes
      summary: Associar plataformas
      description: Associa múltiplas plataformas a um assinante
      operationId: associarPlataformas
      parameters:
      - name: id
        in: path
        description: ID do assinante
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssociarPlataformasRequest'
        required: true
      responses:
        "200":
          description: Associação realizada
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Assinante/Plataforma não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Erro de validação de negócio
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
  /pagamentos/assinante/{id}:
    get:
      tags:
      - Pagamentos
      summary: Listar pagamentos por assinante
      description: Retorna lista paginada de pagamentos de um assinante
      operationId: listarPagamentosAssinante
      parameters:
      - name: id
        in: path
        description: ID do assinante
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: pagina
        in: query
        description: Número da página (0..N)
        required: false
        schema:
          minimum: 0
          type: integer
          format: int32
          default: 0
        example: 0
      - name: quantidade
        in: query
        description: Quantidade por página
        required: false
        schema:
          minimum: 1
          type: integer
          format: int32
          default: 10
        example: 10
      responses:
        "404":
          description: Assinante não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "200":
          description: Lista retornada com sucesso
          content:
            '*/*':
              schema:
                type: object
      security:
      - bearer-key: []
  /assinantes/{assinanteId}/desassociar-plataforma/{plataformaId}:
    delete:
      tags:
      - Assinantes
      summary: Desassociar plataforma
      description: Remove a associação de uma plataforma de um assinante
      operationId: desassociarPlataforma
      parameters:
      - name: assinanteId
        in: path
        description: ID do assinante
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      - name: plataformaId
        in: path
        description: ID da plataforma
        required: true
        schema:
          type: integer
          format: int64
        example: 2
      responses:
        "200":
          description: Desassociação realizada
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Assinante/Plataforma não encontrado
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        "400":
          description: Erro de validação de negócio
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - bearer-key: []
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          description: Data/hora do erro (UTC)
          format: date-time
        status:
          type: integer
          description: Código HTTP do erro
          format: int32
          example: 404
        error:
          type: string
          description: Tipo de erro
          example: VulpesException
        message:
          type: string
          description: Mensagem explicativa do erro
          example: Plataforma não encontrada
        path:
          type: string
          description: Caminho da requisição
          example: /plataformas/999
      description: Estrutura padronizada de erros da API
...
    Assinante:
      required:
      - email
      - nome
      type: object
      properties:
        id:
          type: integer
          description: Identificador do assinante
          format: int64
          example: 1
        nome:
          type: string
          description: Nome do assinante
          example: João Silva
        email:
          type: string
          description: Email do assinante
          example: joao@email.com
        plataformas_associadas:
          type: array
          description: Plataformas associadas ao assinante
          items:
            $ref: '#/components/schemas/PlataformaResumo'
        valor_por_mes:
          type: number
          description: Custo mensal estimado
          example: 29.9
        cadastrado_em:
          type: string
          description: Data de criação
          format: date-time
        atualizado_em:
          type: string
          description: Data de atualização
          format: date-time
        excluido_em:
          type: string
          description: Data de exclusão (soft delete)
          format: date-time
      description: Dados de um assinante
    PlataformaResumo:
      type: object
      properties:
        id:
          type: integer
          description: ID da plataforma
          format: int64
          example: 1
        nome:
          type: string
          description: Nome da plataforma
          example: Netflix
        precoMensal:
          type: number
          description: Preço mensal total
          example: 29.9
        precoIndividual:
          type: number
          description: Preço individual por assinante
          example: 9.97
      description: Resumo de plataforma associada ao assinante
...
    AssociarPlataformasRequest:
      type: object
      properties:
        plataformaIds:
          type: array
          description: IDs das plataformas
          example:
          - 1
          - 2
          - 3
          items:
            type: integer
            description: IDs das plataformas
            format: int64
      description: Requisição para associar plataformas a um assinante
  securitySchemes:
    bearer-key:
      type: http
      scheme: bearer
      bearerFormat: JWT